Reorganize ref and integer alias headers
