+++ Engine/Include/Tbx/Debug/ILogger.h
@@ -2,7 +2,7 @@
 #include "Tbx/DllExport.h"
 #include <string>
 #include <memory>
-#include "Tbx/TypeAliases/Pointers.h"
+#include "Tbx/Memory/Refs/Refs.h"
 
 namespace Tbx
 {
