+++ Engine/Include/Tbx/Events/PluginEvents.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "Tbx/Events/Event.h"
 #include "Tbx/Plugins/LoadedPlugin.h"
-#include "Tbx/TypeAliases/Pointers.h"
+#include "Tbx/Memory/Refs/Refs.h"
 
 namespace Tbx
 {
