+++ Engine/Include/Tbx/Math/Int.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include <cstddef>
+#include <cstdint>
+
+namespace Tbx
+{
+    using uint = unsigned int;
+    using uint32 = std::uint32_t;
+    using uint64 = unsigned long long;
+    using int64 = long long;
+
+    using hash = std::size_t;
+    using byte = unsigned char;
+}
