+++ Engine/Include/Tbx/Plugins/LoadedPlugin.h
@@ -4,7 +4,7 @@
 #include "Tbx/Plugins/SharedLibrary.h"
 #include "Tbx/DllExport.h"
 #include <memory>
-#include "Tbx/TypeAliases/Pointers.h"
+#include "Tbx/Memory/Refs/Refs.h"
 
 namespace Tbx
 {
