--- Engine/Include/Tbx/TypeAliases/Int.h
@@ -1,12 +1,0 @@
-#pragma once
-#include <cstdint>
-
-namespace Tbx
-{
-	using uint = unsigned int;
-	using uint32 = uint32_t;
-	using uint64 = unsigned long long;
-	using int64 = long long;
-    using hash = std::size_t;
-    using byte = unsigned char;
-}
