+++ Engine/Source/Tbx/Math/Mat4x4.cpp
@@ -2,7 +2,7 @@
 #include "Tbx/Math/Constants.h"
 #include "Tbx/Math/Mat4x4.h"
 #include "Tbx/Math/Trig.h"
-#include "Tbx/TypeAliases/Int.h"
+#include "Tbx/Math/Int.h"
 #include <glm/gtc/matrix_transform.hpp>
 #include <glm/gtc/type_ptr.hpp>
 #include <glm/gtx/quaternion.hpp>
